Refactor to abstract away all console commands to a encapsulated class
