+++ tests/test_integration/test_cbz_downloading.py
@@ -2,8 +2,8 @@ import os
 from unittest import mock
 
 
-@mock.patch("cbz_tagger.database.entity_db.get_input")
-@mock.patch("cbz_tagger.database.entity_db.get_raw_input")
+@mock.patch("cbz_tagger.common.input.get_input")
+@mock.patch("cbz_tagger.common.input.get_raw_input")
 @mock.patch("cbz_tagger.entities.chapter_entity.ChapterEntity.download_file")
 def test_download_cbz_files_with_mark_all_tracked(
     mock_download_file, mock_get_raw_input, mock_get_input, integration_scanner, manga_name, capture_input_fixture
@@ -38,8 +38,8 @@ def test_download_cbz_files_with_mark_all_tracked(
     assert len(os.listdir(integration_scanner.scan_path)) == 0
 
 
-@mock.patch("cbz_tagger.database.entity_db.get_input")
-@mock.patch("cbz_tagger.database.entity_db.get_raw_input")
+@mock.patch("cbz_tagger.common.input.get_input")
+@mock.patch("cbz_tagger.common.input.get_raw_input")
 @mock.patch("cbz_tagger.entities.chapter_entity.ChapterEntity.download_file")
 def test_download_cbz_files_without_mark_all_tracked(
     mock_download_file, mock_get_raw_input, mock_get_input, integration_scanner, manga_name, capture_input_fixture
