+++ tests/test_unit/test_common/test_get_input.py
@@ -1,5 +1,5 @@
-from cbz_tagger.common.helpers import get_input
-from cbz_tagger.common.helpers import get_raw_input
+from cbz_tagger.common.input import get_input
+from cbz_tagger.common.input import get_raw_input
 
 
 def test_valid_input(monkeypatch):
